pyproject.toml
[project]
name = "driver-agency-mapping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gspread>=6.2.1",
    "gspread-dataframe>=4.0.0",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "snowflake-connector-python>=3.15.0",
]
